+++ vlingo-reactive-messaging-patterns/src/main/java/io/vlingo/reactive/messaging/patterns/recipientlist/MountainSuppliesOrderProcessor.java
@@ -13,8 +13,6 @@ import java.util.Map;
 
 import io.vlingo.actors.Actor;
 import io.vlingo.actors.testkit.TestUntil;
-import io.vlingo.reactive.messaging.patterns.recipientlist.PriceQuoteInterest.PriceQuote;
-import io.vlingo.reactive.messaging.patterns.recipientlist.RetailBasket.RetailItem;
 
 /**
  * MountainSuppliesOrderProcessor maintains registry of @{@link QuoteProcessor} {@link Actor} instances
@@ -84,210 +82,4 @@ implements OrderProcessor
             }
         }
     }
-
-    // QUOTE PROCESSORS
-    public static class BudgetHikersPriceQuotes 
-    extends Actor 
-    implements QuoteProcessor
-    {
-        public final OrderProcessor orderProcessor;
-        
-        public BudgetHikersPriceQuotes( final OrderProcessor orderProcessor )
-        {
-            this.orderProcessor = orderProcessor;
-        }
-
-        /* @see io.vlingo.actors.Actor#beforeStart() */
-        @Override
-        protected void beforeStart()
-        {
-            orderProcessor.register( new PriceQuoteInterest( this.getClass().getName(), selfAs( QuoteProcessor.class ), 1d, 1000d ));
-        }
-
-        /* @see io.vlingo.reactive.messaging.patterns.recipientlist.QuoteProcessor#requestPriceQuote(io.vlingo.reactive.messaging.patterns.recipientlist.RetailBasket) */
-        @Override
-        public void requestPriceQuote( String rfqId, String itemId, Double retailPrice, Double totalRetailPrice )
-        {
-            Double discountPercentage = discountPercentage( totalRetailPrice );
-            Double discountPrice = retailPrice - ( retailPrice * discountPercentage );
-            orderProcessor.remittedPriceQuote( new PriceQuote( selfAs( QuoteProcessor.class ), rfqId, itemId, retailPrice, discountPrice ));
-        }
-
-        protected Double discountPercentage( Double orderTotalRetailprice )
-        {
-            Double discount = null;
-            if ( orderTotalRetailprice <= 100.00 ) discount = .02;
-            else if ( orderTotalRetailprice <= 399.99 ) discount = .03;
-            else if ( orderTotalRetailprice <= 499.99 ) discount = .05;
-            else if ( orderTotalRetailprice <= 799.99 ) discount = .07;
-            else discount = .075;
-            return discount;
-        }
-    }
-
-    public static class HighSierraPriceQuotes 
-    extends Actor 
-    implements QuoteProcessor
-    {
-        public final OrderProcessor orderProcessor;
-        
-        public HighSierraPriceQuotes( final OrderProcessor orderProcessor )
-        {
-            this.orderProcessor = orderProcessor;
-        }
-
-        /* @see io.vlingo.actors.Actor#beforeStart() */
-        @Override
-        protected void beforeStart()
-        {
-            orderProcessor.register( new PriceQuoteInterest( this.getClass().getName(), selfAs( QuoteProcessor.class ), 100d, 10000d ));
-        }
-
-        /* @see io.vlingo.reactive.messaging.patterns.recipientlist.QuoteProcessor#requestPriceQuote(io.vlingo.reactive.messaging.patterns.recipientlist.RetailBasket) */
-        @Override
-        public void requestPriceQuote( String rfqId, String itemId, Double retailPrice, Double totalRetailPrice )
-        {
-            Double discountPercentage = discountPercentage( totalRetailPrice );
-            Double discountPrice = retailPrice - ( retailPrice * discountPercentage );
-            orderProcessor.remittedPriceQuote( new PriceQuote( selfAs( QuoteProcessor.class ), rfqId, itemId, retailPrice, discountPrice ));
-        }
-
-        protected Double discountPercentage( Double orderTotalRetailprice )
-        {
-            Double discount = null;
-            if ( orderTotalRetailprice <= 150.00 ) discount = .015;
-            else if ( orderTotalRetailprice <= 499.99 ) discount = .02;
-            else if ( orderTotalRetailprice <= 999.99 ) discount = .03;
-            else if ( orderTotalRetailprice <= 4999.99 ) discount = .04;
-            else discount = .05;
-            return discount;
-        }
-    }
-    
-    public static class MountainAscentPriceQuotes 
-    extends Actor 
-    implements QuoteProcessor
-    {
-        public final OrderProcessor orderProcessor;
-        
-        public MountainAscentPriceQuotes( final OrderProcessor orderProcessor )
-        {
-            this.orderProcessor = orderProcessor;
-        }
-
-        /* @see io.vlingo.actors.Actor#beforeStart() */
-        @Override
-        protected void beforeStart()
-        {
-            orderProcessor.register( new PriceQuoteInterest( getClass().getName(), selfAs( QuoteProcessor.class ), 70d, 500d ));
-        }
-
-        /* @see io.vlingo.reactive.messaging.patterns.recipientlist.QuoteProcessor#requestPriceQuote(io.vlingo.reactive.messaging.patterns.recipientlist.RetailBasket) */
-        @Override
-        public void requestPriceQuote( String rfqId, String itemId, Double retailPrice, Double totalRetailPrice )
-        {
-            Double discountPercentage = discountPercentage( totalRetailPrice );
-            Double discountPrice = retailPrice - ( retailPrice * discountPercentage );
-            orderProcessor.remittedPriceQuote( new PriceQuote( selfAs( QuoteProcessor.class ), rfqId, itemId, retailPrice, discountPrice ));
-        }
-
-        protected Double discountPercentage( Double orderTotalRetailprice )
-        {
-            Double discount = null;
-            if ( orderTotalRetailprice <= 99.99 ) discount = .01;
-            else if ( orderTotalRetailprice <= 199.99 ) discount = .02;
-            else if ( orderTotalRetailprice <= 499.99 ) discount = .03;
-            else if ( orderTotalRetailprice <= 799.99 ) discount = .04;
-            else if ( orderTotalRetailprice <= 999.99 ) discount = .045;
-            else if ( orderTotalRetailprice <= 2999.99 ) discount = .0475;
-            else discount = .05;
-            return discount;
-        }
-    }
-    
-    public static class PinnacleGearPriceQuotes 
-    extends Actor 
-    implements QuoteProcessor
-    {
-        public final OrderProcessor orderProcessor;
-        
-        public PinnacleGearPriceQuotes( final OrderProcessor orderProcessor )
-        {
-            this.orderProcessor = orderProcessor;
-        }
-
-        /* @see io.vlingo.actors.Actor#beforeStart() */
-        @Override
-        protected void beforeStart()
-        {
-            orderProcessor.register( new PriceQuoteInterest( getClass().getName(), selfAs( QuoteProcessor.class ), 250d, 500000d ));
-        }
-
-        /* @see io.vlingo.reactive.messaging.patterns.recipientlist.QuoteProcessor#requestPriceQuote(io.vlingo.reactive.messaging.patterns.recipientlist.RetailBasket) */
-        @Override
-        public void requestPriceQuote( String rfqId, String itemId, Double retailPrice, Double totalRetailPrice )
-        {
-            Double discountPercentage = discountPercentage( totalRetailPrice );
-            Double discountPrice = retailPrice - ( retailPrice * discountPercentage );
-            orderProcessor.remittedPriceQuote( new PriceQuote( selfAs( QuoteProcessor.class ), rfqId, itemId, retailPrice, discountPrice ));
-        }
-
-        protected Double discountPercentage( Double orderTotalRetailprice )
-        {
-            Double discount = null;
-            if ( orderTotalRetailprice <= 299.99 ) discount = .015;
-            else if ( orderTotalRetailprice <= 399.99 ) discount = .0175;
-            else if ( orderTotalRetailprice <= 499.99 ) discount = .02;
-            else if ( orderTotalRetailprice <= 999.99 ) discount = .03;
-            else if ( orderTotalRetailprice <= 1199.99 ) discount = .035;
-            else if ( orderTotalRetailprice <= 4999.99 ) discount = 0.04;
-            else if ( orderTotalRetailprice <= 7999.99 ) discount = 0.05;
-            else discount = .06;
-            return discount;
-        }
-    }
-
-    public static class RockBottomOuterwearPriceQuotes 
-    extends Actor 
-    implements QuoteProcessor
-    {
-        public final OrderProcessor orderProcessor;
-        
-        public RockBottomOuterwearPriceQuotes( final OrderProcessor orderProcessor )
-        {
-            this.orderProcessor = orderProcessor;
-        }
-        
-        /* @see io.vlingo.actors.Actor#beforeStart() */
-        @Override
-        protected void beforeStart()
-        {
-            orderProcessor.register( new PriceQuoteInterest( getClass().getName(), selfAs( QuoteProcessor.class ), .50, 7500d ));
-        }
-
-        /* @see io.vlingo.reactive.messaging.patterns.recipientlist.QuoteProcessor#requestPriceQuote(io.vlingo.reactive.messaging.patterns.recipientlist.RetailBasket) */
-        @Override
-        public void requestPriceQuote( String rfqId, String itemId, Double retailPrice, Double totalRetailPrice )
-        {
-            Double discountPercentage = discountPercentage( totalRetailPrice );
-            Double discountPrice = retailPrice - ( retailPrice * discountPercentage );
-            orderProcessor.remittedPriceQuote( new PriceQuote( selfAs( QuoteProcessor.class ), rfqId, itemId, retailPrice, discountPrice ));
-        }
-
-        protected Double discountPercentage( Double orderTotalRetailprice )
-        {
-            Double discount = null;
-            if ( orderTotalRetailprice <= 100.00 ) discount = .015;
-            else if ( orderTotalRetailprice <= 399.99 ) discount = .02;
-            else if ( orderTotalRetailprice <= 499.99 ) discount = .03;
-            else if ( orderTotalRetailprice <= 799.99 ) discount = .04;
-            else if ( orderTotalRetailprice <= 999.99 ) discount = .05;
-            else if ( orderTotalRetailprice <= 2999.99 ) discount = .06;
-            else if ( orderTotalRetailprice <= 4999.99 ) discount = 0.07;
-            else if ( orderTotalRetailprice <= 5999.99 ) discount = 0.075;
-            else discount = .08;
-            return discount;
-        }
-    }
-    
 }
